pyproject.toml
[project]
name = "fast-seeder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "faker>=37.4.0",
    "fastapi[standard]>=0.116.0",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.9",
]

[tool.ruff.lint]
select = ['E', 'F', 'W', 'A', 'PLC', 'PLW', 'I']
ignore = ["A005"]

[tool.ruff]
line-length = 120
exclude = [".venv", ".git", ".vscode"]
